refactor: move util types to separate file
