+++ packages/next-safe-action/src/hooks.ts
@@ -8,8 +8,8 @@ import {} from "react/experimental";
 import type {} from "zod";
 import { getActionShorthandStatusObject, getActionStatus, useActionCallbacks } from "./hooks-utils";
 import type { HookCallbacks, HookResult, HookSafeActionFn } from "./hooks.types";
-import type { InferIn, Schema } from "./utils";
 import { isError } from "./utils";
+import type { InferIn, Schema } from "./utils.types";
 
 // HOOKS
