+++ packages/next-safe-action/src/validation-errors.types.ts
@@ -1,4 +1,4 @@
-import type { Infer, Prettify, Schema } from "./utils";
+import type { Infer, Prettify, Schema } from "./utils.types";
 
 export type ValidationIssue = {
 	message: string;
